Create footer component
