+++ src/components/footer/footer.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Footer = () => (
+  <footer>
+    <div>
+      <p>Copyright © Damilola Ale 2020</p>
+    </div>
+  </footer>
+);
+
+export default Footer;
